**Description:**
Adds the federal account code to the response from v2/federal_obligations

**Technical details:**
Adds the account code as an annotation to the queryset. 

**Requirements for PR merge:**

1. [X] Unit & integration tests updated
2. [NA] API documentation updated 
3. [x] Necessary PR reviewers:
	- [x] Backend
	- [x] Frontend (Lizzie)
4. [NA] Matview impact assessment completed
5. [X] Frontend impact assessment completed
6. [X] Data validation completed
8. [X] Jira Ticket [DEV-1552](https://federal-spending-transparency.atlassian.net/browse/DEV-1552):
	- [x] Link to this Pull-Request
	- [x] Performance evaluation of affected (API | Script | Download)
	- [x] Before / After data comparison

**Area for explaining above N/A when needed:**
```
Frontend will be updating the API Contract
```

Performance:
Before:
1248 ms
1337 ms
1283 ms
1278 ms
1316 ms
1298 ms
1249 ms


After
1275 ms
1293 ms
1260 ms
1296 ms
1286 ms
1281 ms
1279 ms

No Significant Change